Make search paths configurable
The current hard-coded search paths only work for a particular
Cucumber setup.

Created a configuration setting to allow customizing the paths.

Defaulted to searching through all *.js and *.rb files, which
should work for most users.  It might not be fast enough, in which
case a more specific path can be configured.